% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSignatureDatabase.R
\name{create_tcga_pancan_database}
\alias{create_tcga_pancan_database}
\title{Create TCGA pan-cancer-atlas signature database}
\usage{
create_tcga_pancan_database(outdir = getwd())
}
\arguments{
\item{outdir}{output directory}
}
\value{
run for its side-effects. Invisibly returns NULL
}
\description{
Creates an sqlite database describing the mutational signatures present in every TCGA patient
}
\examples{
\dontrun({create_tcga_pancan_database})
}
