% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSignatureDatabase.R
\name{tcga_fetch_metadata}
\alias{tcga_fetch_metadata}
\title{Fetch TCGA metadata}
\usage{
tcga_fetch_metadata(ids = NULL)
}
\value{
data.frame with TCGA metadata
}
\description{
Fetch TCGA metadata
}
