% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSignatureDatabase.R
\name{tcga_fetch_pancan_maf}
\alias{tcga_fetch_pancan_maf}
\title{Fetch MAF}
\usage{
tcga_fetch_pancan_maf(ids = NULL)
}
\arguments{
\item{ids}{TCGA IDs. See \code{\link[=tcga_fetch_metadata]{tcga_fetch_metadata()}} for a full list of possible IDs. Set to NULL to gt all data}
}
\value{
maf data.frame
}
\description{
Fetch TCGA MC3 MAF
}
