% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSignatureDatabase.R
\name{tcga_fetch_metadata}
\alias{tcga_fetch_metadata}
\title{Fetch TCGA metadata}
\usage{
tcga_fetch_metadata(ids = NULL)
}
\arguments{
\item{ids}{a character vector of TCGA IDs to analyse. If NULL will run on all TCGA samples. See \code{\link[=tcga_fetch_metadata]{tcga_fetch_metadata()}} for a full list of possible IDs}
}
\value{
data.frame with TCGA metadata
}
\description{
Fetch TCGA metadata
}
